:recycle: [Refactor] comment header, item feature의ui로 이동 #main
